Remove 'chromedriver' from dependencies
Please note that it is just a (rather broken) Python package.
Its purpose is solely to download a `chromedriver` binary
from the Google servers and make it discoverable by simply importing
the module.

The reason for removal is that it currently breaks the build

If you wish to use the test suites with the Chrome / Chromium browser,
you should manually install a `chromedriver` binary of an appropriate
version into one of your `$PATH` directories.